getLong was using INT. Patch by Criag Russell.
git-svn-id: https://svn.apache.org/repos/asf/openjpa/trunk@808701 13f79535-47bb-0310-9956-ffa450edef68

+++ openjpa-jdbc/src/main/java/org/apache/openjpa/jdbc/sql/AbstractResult.java
@@ -661,8 +661,8 @@ public abstract class AbstractResult
     protected long getLongInternal(Object obj, Joins joins)
         throws SQLException {
         Number val = (Number) checkNull(getObjectInternal(obj,
-            JavaTypes.INT, null, joins));
-        return (val == null) ? 0 : val.intValue();
+            JavaTypes.LONG, null, joins));
+        return (val == null) ? 0 : val.longValue();
     }
 
     public Number getNumber(Object obj)
